security contract transfered to securityLib

+++ contracts/libraries/VoteLib.sol
@@ -3,6 +3,7 @@ pragma abicoder v2;
 
 import "./PostLib.sol";
 
+
 /// @title VoteLib
 /// @notice Provides information about operation with posts                     //
 /// @dev posts information is stored in the mapping on the main contract        ///
@@ -10,51 +11,51 @@ library VoteLib  {
     enum ResourceAction { Downvote, Upvoted, Downvoted, AcceptReply, AcceptedReply, FirstReply, QuickReply }
 
     //expert post
-    int8 constant DownvoteExpertPost = -1;
-    int8 constant UpvotedExpertPost = 5;
-    int8 constant DownvotedExpertPost = -2;
-    int8 constant AcceptExpertPost = 2;         //Accept answer as correct for Expert Question
+    int32 constant DownvoteExpertPost = -1;
+    int32 constant UpvotedExpertPost = 5;
+    int32 constant DownvotedExpertPost = -2;
+    int32 constant AcceptExpertPost = 2;         //Accept answer as correct for Expert Question
 
     //common post 
-    int8 constant DownvoteCommonPost = -1;
-    int8 constant UpvotedCommonPost = 1;
-    int8 constant DownvotedCommonPost = -1;
-    int8 constant AcceptCommonPost = 1;
+    int32 constant DownvoteCommonPost = -1;
+    int32 constant UpvotedCommonPost = 1;
+    int32 constant DownvotedCommonPost = -1;
+    int32 constant AcceptCommonPost = 1;
 
     //tutorial 
-    int8 constant DownvoteTutorial = -1;
-    int8 constant UpvotedTutorial = 1;
-    int8 constant DownvotedTutorial = -1 ;
+    int32 constant DownvoteTutorial = -1;
+    int32 constant UpvotedTutorial = 1;
+    int32 constant DownvotedTutorial = -1 ;
 
-    int8 constant DeleteOwnPost = -1;
-    int8 constant ModeratorDeletePost = -2;
+    int32 constant DeleteOwnPost = -1;
+    int32 constant ModeratorDeletePost = -2;
 
 /////////////////////////////////////////////////////////////////////////////
 
     //expert reply
-    int8 constant DownvoteExpertReply = -1;
-    int8 constant UpvotedExpertReply = 10;
-    int8 constant DownvotedExpertReply = -2;
-    int8 constant AcceptExpertReply = 15;
-    int8 constant AcceptedExpertReply = 2;
-    int8 constant FirstExpertReply = 5;
-    int8 constant QuickExpertReply = 5;
+    int32 constant DownvoteExpertReply = -1;
+    int32 constant UpvotedExpertReply = 10;
+    int32 constant DownvotedExpertReply = -2;
+    int32 constant AcceptExpertReply = 15;
+    int32 constant AcceptedExpertReply = 2;
+    int32 constant FirstExpertReply = 5;
+    int32 constant QuickExpertReply = 5;
 
     //common reply 
-    int8 constant DownvoteCommonReply = -1;
-    int8 constant UpvotedCommonReply = 2;
-    int8 constant DownvotedCommonReply = -1;
-    int8 constant AcceptCommonReply = 3;
-    int8 constant AcceptedCommonReply = 1;
-    int8 constant FirstCommonReply = 1;
-    int8 constant QuickCommonReply = 1;
+    int32 constant DownvoteCommonReply = -1;
+    int32 constant UpvotedCommonReply = 2;
+    int32 constant DownvotedCommonReply = -1;
+    int32 constant AcceptCommonReply = 3;
+    int32 constant AcceptedCommonReply = 1;
+    int32 constant FirstCommonReply = 1;
+    int32 constant QuickCommonReply = 1;
     
-    int8 constant DeleteOwnReply = -1;
-    int8 constant ModeratorDeleteReply = -2;            // to do
+    int32 constant DeleteOwnReply = -1;
+    int32 constant ModeratorDeleteReply = -2;            // to do
 
 /////////////////////////////////////////////////////////////////////////////////
 
-    int8 constant ModeratorDeleteComment = -1;
+    int32 constant ModeratorDeleteComment = -1;
 
     /// @notice Get value Rating for post action
     /// @param postType Type post: expertPost, commonPost, tutorial
@@ -62,7 +63,7 @@ library VoteLib  {
     function getUserRatingChangeForPostAction(
         PostLib.PostType postType,
         ResourceAction resourceAction
-    ) internal pure returns (int8) {
+    ) internal pure returns (int32) {
  
         if (PostLib.PostType.ExpertPost == postType) {          //switch, gas?
             if (ResourceAction.Downvote == resourceAction) return DownvoteExpertPost;
@@ -90,7 +91,7 @@ library VoteLib  {
     function getUserRatingChangeForReplyAction(
         PostLib.PostType postType,
         ResourceAction resourceAction
-    ) internal pure returns (int8) {
+    ) internal pure returns (int32) {
  
         if (PostLib.PostType.ExpertPost == postType) {          //switch, gas?
             if (ResourceAction.Downvote == resourceAction) return DownvoteExpertReply;
@@ -122,7 +123,7 @@ library VoteLib  {
         PostLib.PostType postType,
         ResourceAction resourceAction,
         PostLib.TypeContent typeContent
-    ) internal pure returns (int8) {
+    ) internal pure returns (int32) {
         if (PostLib.TypeContent.Post == typeContent) {
             return getUserRatingChangeForPostAction(postType, resourceAction);
         } else if (PostLib.TypeContent.Reply == typeContent) {
@@ -146,9 +147,9 @@ library VoteLib  {
         mapping(address => int256) storage historyVotes,
         bool isUpvote,
         address[] storage votedUsers                              /// for comment
-    ) internal returns (int8) {
+    ) internal returns (int32) {
         int history = getHistoryVote(actionAddress, historyVotes);
-        int8 ratingChange;
+        int32 ratingChange;
         
         if (isUpvote) {
             if (history == -1) {
